backend/requirements.txt
aniso8601==9.0.1
Babel==2.14.0
blinker==1.7.0
click==8.1.7
dnspython==2.6.1
email_validator==2.1.1
Flask==3.0.2
Flask-BabelEx==0.9.4
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security==3.0.0
Flask-Security-Too==5.4.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
idna==3.6
importlib_resources==6.1.3
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
passlib==1.7.4
pillow==10.2.0
pytz==2024.1
six==1.16.0
speaklater==1.3
SQLAlchemy==2.0.28
typing_extensions==4.10.0
Werkzeug==3.0.1
WTForms==3.1.2
